+++ src/bielsort/__init__.py
@@ -13,15 +13,21 @@ from bielsort_native import (
     biel_sort_in_place_with_strategy,
     biel_sort_with_strategy,
     sort,
+    sort_in_place,
+    sort_in_place_with_strategy,
+    sort_with_strategy,
 )
 
 __all__ = [
+    "sort",
+    "sort_with_strategy",
+    "sort_in_place",
+    "sort_in_place_with_strategy",
     "biel_sort",
     "biel_sort_diagnostico",
     "biel_sort_with_strategy",
     "biel_sort_in_place",
     "biel_sort_in_place_diagnostico",
     "biel_sort_in_place_with_strategy",
-    "sort",
     "__version__",
 ]